+++ app/controllers/admin/users_controller.rb
@@ -10,7 +10,7 @@ module Admin
     private
 
     def users(query)
-      User.filter_by_email(query)
+      User.search(query)
         .includes(developer: {avatar_attachment: :blob}, business: {avatar_attachment: :blob})
         .order(:email)
     end